MINOR: Add test case for duplicate check after log cleaning

+++ clients/src/main/java/org/apache/kafka/common/record/DefaultRecord.java
@@ -102,7 +102,7 @@ public class DefaultRecord implements Record {
     }
 
     @Override
-    public long sequence() {
+    public int sequence() {
         return sequence;
     }
